reorganize ui code
